Rename NativeLibraryFromPrebuiltAar to PrebuiltNativeLibraryDir
Summary:
I'm working on implementing `prebuilt_native_library`. Its bundling
logic is identical to that for the native libraries in a prebuilt AAR,
so I'm just going to reuse that, but it felt strange to keep the old
name in that case. I didn't want to rename it to just PrebuiltNativeLibrary
though, since that's too easily confused with the rule. After lots of
agonizing (naming things is hard), I settled on PrebuiltNativeLibraryDir
to indicate a directory of prebuilt native libraries, which is accurate
for both the prebuilt AAR and the `prebuilt_native_library` case. I also
renamed all related attributes:

* NativeLibraryFromPrebuiltAar -> PrebuiltNativeLibraryDir
* NativeLibraryFromPrebuiltAarTSet -> PrebuiltNativeLibraryDirTSet
* native_libs_from_prebuilt_aars_to_exclude -> prebuilt_native_library_dirs_to_exclude
* native_lib_from_prebuilt_aar -> prebuilt_native_library_dir
* native_libs_from_prebuilt_aars -> prebuilt_native_library_dirs
* apk_under_test_native_libs_from_prebuilt_aars -> apk_under_test_prebuilt_native_library_dirs
* traversed_native_libs_from_prebuilt_aars -> traversed_prebuilt_native_library_dirs
* all_native_libs_from_prebuilt_aars -> all_prebuilt_native_library_dirs
* native_libs_from_prebuilt_aars_for_system_library_loader -> prebuilt_native_library_dirs_for_system_library_loader

I didn't rename the shorter variable names used in various places (e.g.
`native_library`, `native_lib`, or `native_libs`), for brevity and
because they didn't mention anything about AARs anyway.

This is a pure renaming change, and it's separate out from the actual
`prebuilt_native_library` implementation for ease of review.

Reviewed By: IanChilds

Differential Revision: D38230274

fbshipit-source-id: b66b8af497fd75c2cf2b25fd9400482a095bbff9