buck2: re: uploader: allow uploading relative to a path
Summary:
Like it says in the title. I want to reuse this for uploading action outputs.
One difference between inputs and outputs is that the former are rooted at the
project root but the latter are not.

So, let's extend the Uploader to allow uploading a directory that is *not*
relative to the project root. All we need to do in order to achieve this is
prefix the paths that we give the materializer and / or the RE client.

Note that of course in the context of uploading action outputs we won't go
to the materializer, but while that's a bit of unnecessary overhead, I doubt
it's worth worrying about.

Reviewed By: bobyangyf

Differential Revision: D38212395

fbshipit-source-id: cad910814d9b0f573a14281d28c05272e143e9c1