buck2: forkserver: refactor gather_output
Summary:
Rework this function as an explicit stream instead of a future.

This is necessary because right now the reason we can make it a future
if because we buffer all the outputs, but in the fork server we don't want to do
that because our goal is to minimize our memory footprint (so we want the daemon
to be the one to buffer).

Not much is actually changing here. Instead of joining two futures, we write a
stream that returns events whenever we get something on stdout, and ensures
we return the status last (this "ensure we return the status last" is why we need
a custom stream as opposed to just `futures::stream::select`).

Reviewed By: bobyangyf

Differential Revision: D38123839

fbshipit-source-id: 5a8efefd831525278b915467e9e0c9b4a7bf92ff