buck2: local executor: rework the "read ready" logic
Summary:
Like it says in the title. This updates our "read ready" logic (we use this
so that if th FD we passed our child was sent somewhere else we don't block
on that) to be represented as an AsyncRead rather than as future returning
a `Vec<u8>` (this was the `fn drain()` this diff removes).

I need to do this because I want to reuse this in Forkserver, but I don't
want to do any buffering there.

A nice side effect of this is that we are decoupling two bits of logic here,
which means we can more easily test it:

- Switching to draining (note: I'm calling "draining" the process of reading
  what is ready but not waiting for further data) when we are interrupted;
- Actually doing the draining.

Note that the latter is platform specific (on Unix we do a non-blocking
read, on Windows we just read with a 1 second timeout), so this lets us
cover more things (in fact, all things) with tests.

Reviewed By: ndmitchell

Differential Revision: D38116130

fbshipit-source-id: 77a09ef31d21a9bcefb738785035254751851a29