Migrate get_bound in InstallInfo to get_bound_artifact
Summary:
Please see D37883940 (https://github.com/facebookincubator/buck2/commit/69f0fec47fbbc6e917afbd4e113b47bc31c1c306) for a bit more context

The InstallInfo provider provides a simple interface: a RunInfo and a set of files to be installed. As the user controls the set of files and how they are processed, it would be surprising for there to be additional files that pop up.

Maybe we could provide an option to expand each file into any additional inputs that exists in the future, but I'm not sure if that's a behavior we currently need (from a quick glance at how the provider is used).

This is a case where we do not need all the additional input files we need for build here. We can migrate to get_bound_artifact which is a logical noop.

Reviewed By: krallin

Differential Revision: D37937619

fbshipit-source-id: f78f5753774b45f3c9c61f1889546c728d58c793