Migrate default info iterations to also look at additional_artifacts
Summary:
Migrate handling of default / other outputs to also consider additional inputs tracked in Starlark*Artifact.

Essentially, whatever operation we are applying to other outputs should be applied to the additional artifacts as well.

After the migrations in this stack is done, we can go through rule by rule so we can deprecate other outputs.

Reviewed By: bobyangyf

Differential Revision: D38132475

fbshipit-source-id: b403f245c6d61154ae8c1b9036e70be6fb3f7ea1