Improve error messaging around unimplemented function (labels)
Summary:
Temporary improvement until I get around to actually implementing the labels functionality. The current error you get is misleading, we should properly tell the user that it's not implemented yet.

The function signature should match the intended signature even if it is yet to be implemented.

Might be good to a survey of any other unimplemented functions to see if it is behaving this way.

Reviewed By: bobyangyf

Differential Revision: D38249577

fbshipit-source-id: 03c52a71aa9337ff5983ebc7dfa12d08e46add48