Add a dedupe_by_value utility function
Summary:
Until we have this in the core (see T127134666), just write it in Starlark. Open to ideas of how to make it most performant etc, though hopefully it won't be needed for long.

This is needed because e.g. `raw_target()` returns a new object each time, so pointer equality does not work.

Reviewed By: bobyangyf

Differential Revision: D38242400

fbshipit-source-id: 5da85d83187c161dd13082bfd0723a53a713fdb5