Move link groups code from apple to cxx
Summary:
Link groups are used in both apple and cxx rules, and should live in the cxx rules as apple uses cxx, not the other way around.

Cxx should not depend on Apple. Link groups used to be the only violator, but another has been added. I'll clean that up later.

Reviewed By: milend

Differential Revision: D38218046

fbshipit-source-id: 2eb5335281c8e1dc37f83ba7deb5054aa3c4ebd8