buck2: cli: add forkserver command
Summary:
Like it says in the title. We need to run and distribute a Forkserver. We have no intention of
handing this over to another team, so let's simply package it in the Buck binary.

Note that we run the Forkserver on a single thread. I actually think this is sufficient for what it
does but I might revisit this.

Reviewed By: stepancheg

Differential Revision: D38200695

fbshipit-source-id: e0025ec7891f649b8468f7cbb1d524b1b98f765d