Cache groups mappings computation
Summary:
One of the most computationally costly of link groups is computing the groups mappings (also the case for resource groups).

Buck1 [caches these by the link group map](https://www.internalfb.com/code/fbsource/[9f82d4285f420a15c79984d80e5ae7575ad57464]/xplat/build_infra/buck_client/src/com/facebook/buck/cxx/LinkableListFilterFactory.java?lines=49-50). This is a bit more complex to do in buck2 because
1. targets don't share resources, and the bottom up analysis requires passing up any relevant data
2. starlark data structures are primitive

We can solve (or at least improve) 1 by wrapping the mapping result in a provider and returning it for a node in the graph. Typically link groups depend on other link groups, meaning the first link group to compute the mapping can provide it up to the next link group to need it.

For 2, we convert the mapping to a string which we then use for comparison. This allows us to verify that the link group map is the same and reuse the result provided by 1.

NOTE: In this implementation, we currently error if the map is not the same. Typically, in all apps that currently use link (and resource) groups, this map is the same, and needs to be for link correctness. The link group spec in itself doesn't require it, but until users seek to use different maps, let's enforce this.

Testing performance with:
```
buck2 kill
buck2 audit providers APP
```

Before and after measurements for the following targets on an M1 Max MBP:
```
igios - 86.7 to 60.4s - 1.4x improvement
fbios-no-extensions - 887.1 to 352.5 - 2.5x improvement
messenger-ios - 52.6 to 51.1 - within margin of error
```

Reviewed By: milend

Differential Revision: D38218045

fbshipit-source-id: 97bc2fa81899cc8cbe31e16de125a593d83b347f