Update prebuilt native library packaging
Summary:
Instead of creating dummy directories for each ABI when unpacking a
prebuilt AAR (if the AAR doesn't have native libraries for that ABI),
use the directory as-is, and add a script for ABI filtering to handle
potentially missing directories (as suggested by IanChilds in D38234535).
This will be used to implemented `prebuilt_native_library` in the next
diff.

#forcetdhashing

Reviewed By: IanChilds

Differential Revision: D38266671

fbshipit-source-id: 089ce24bc754f5a14ab98e9a0cc06e6a4ec6e432