Add "allbuildfiles(...)" to buck2 query
Summary:
Following Bazel's implementation at https://bazel.build/reference/query#buildfiles, we implement the ```allbuildfiles(...)``` query operator. This operator takes in a target (or set of targets) and, prints its corresponding buildfile, any .bzl files it references via load or transitive load.

Note: ```allbuildfiles()``` runs on the unconfigured target graph with buck2 uquery. If invoked as a cquery command, the result is identical to that of uquery.

Reviewed By: stepancheg

Differential Revision: D37800960

fbshipit-source-id: f1ee4279e8cfa3831af1158b7c8d125ef15ec0bd