Simplify link/resource groups mapping computation caching
Summary:
When we cached the computations, we converted the groups to a string and compared them. This required passing the groups in the group info provider.

This isn't exactly necessary. Instead, we can hash the string and just pass the hash. This simplifies comparison and can improve performance slightly.

For `fbios-no-extensions, this saves about 5 to 7 more seconds.

Reviewed By: luciang

Differential Revision: D38304501

fbshipit-source-id: 67ca2b3077fd41e4684b3be79cb5e8dff36840d0