Add a new function get_artifact_path to StarlarkArtifactLike
Summary:
fingerprint, as far as I can see, should be something like a key or a hash that can be used to check for equality between to starlark artifacts. The current implementation uses the artifact path, but we will need to include some information about the associated artifacts in the future.

There is one usage of fingerprint() in artifact/fs.rs that uses fingerprint to actually get the artifact path. This diff adds a separate function so that we don't have this implicit assumption.

Reviewed By: bobyangyf

Differential Revision: D38364859

fbshipit-source-id: 3d6e564b40c5a20ba2d2d4cb178cc4b9ff8dd5f6