Collect dependencies of each transitive dependency
Summary:
For each target, we save its raw target label and the raw target label of each of its deps.

By doing so, we will be able to "reconstruct" the build graph.

Note that there are rules that are used in the Android build graph (e.g. `cxx_library`) that currently do not merge the `AndroidPackageableInfo`), that will be added in future diffs too.

Reviewed By: mykola-semko

Differential Revision: D38351314

fbshipit-source-id: a52b90a91be7bf8076c952ad4a3addff820136c0