buck2: cxx: add [shared-libraries] sub-target to help automatic packaging of dynamic linked binaries and dependencies
Summary:
We need to be able to discover the binary and all of its runtime dependencies.

For shared libraries we need to know the exact path where the runtime dependencies are placed in as that path is embedded in the binary as RPATH

```
$ readelf -d path/to/dynamic-linked-binary | grep RPATH
 0x000000000000000f (RPATH)              Library rpath: [$ORIGIN/__shared_libs_symlink_tree]
```

We also need to know which are the shared libraries that the binary needs and possibly their dwp - in case they were built with Split DWARF and we want to package debug info in a fbpkg.

Add a helper json "index" that points to the extra targets that need to be built during fbpkg building to get the output paths to these:
- path/to:binary[shared-libraries] has an index of other sub-targets to build: libraries, librariesdwp and rpath_tree
- sub-targets for all of the libraries + the rpath-tree

(See integration with fbpkg.builder in {D38337622})

Reviewed By: christycylee

Differential Revision: D38377883

fbshipit-source-id: 29496b604920ab890f011f0d615a9ea0d1f12262