Move preferred_linkage to cxx_context
Summary:
This is part of a series that tries to lift 'ctx' out of
the core cxx UDRs. The goal is to enable other UDRs to re-use these functions
in the context of another rule, e.g. dynamically declaring a cpp_library as
part of the UDR implementation of thrift/protobuf, or making omnibus easier to
use from new languages. My specific goal is to convert a LinkGroup definition map
to a series of cxx_library() invocations from within the cxx_binary() implemenation.
Right now there's no way to do this dynamic declaration because all functions accept ctx
and use it to access cxx.attrs which is a "magic bag" of attributes. Not only is this weakly
typed, but it's also imposible to construct, so when all you have is the ctx for one rule,
there's no way to make it behave like the ctx for another rule.
The eventual goal will be to have ctx exist on the "edges" (top level _impl()
functions) and to be converted to cxx_context as soon as possible. This will provide 2 benefits:
the attr bundles that will exist inside cxx_context will be strongly typed and documented
CxxContext is a constructible type, unlike ctx, so UDR implementations will be able to make
one of their own to pass into prelude/cxx/* making these rule definitions much more re-usable.

This particular diff moves the `ctx.attrs.preferred_linkage` to
`CxxContext`

Reviewed By: christycylee

Differential Revision: D38357479

fbshipit-source-id: dd1647e1d5070be956bc0be24a555a4a9ebe6b41