Start refactoring ctx -> param bundles
Summary:
This is the first diff in a series that tries to lift 'ctx' out of
the core cxx UDRs. The goal is to enable other UDRs to re-use these functions
in the context of another rule, e.g. dynamically declaring a cpp_library as
part of the UDR implementation of thrift/protobuf, or making omnibus easier to
use from new languages. My specific goal is to convert a LinkGroup definition map
to a series of cxx_library() invocations from within the cxx_binary() implemenation.

Right now there's no way to do this dynamic declaration because all functions accept ctx
and use it to access cxx.attrs which is a "magic bag" of attributes. Not only is this weakly
typed, but it's also imposible to construct, so when all you have is the ctx for one rule,
there's no way to make it behave like the ctx for another rule.

This first diff in the series makes things worse, not better :O It adds a cxx_context
which encapsulates the most common usages for cxx outside of attr access: actions, label, and
toolchain info. It threads this new context into all the functions, with the goal of eventually
attaching attribute values to this record so that ctx will no longer be needed in as many of these
functions as possible. The eventual goal will be to have ctx exist on the "edges" (top level _impl()
functions) and to be converted to cxx_context as soon as possible. This will provide 2 benefits:
1) the attr bundles that will exist inside cxx_context will be strongly typed and documented
2) CxxContext is a constructible type, unlike ctx, so UDR implementations will be able to make
   one of their own to pass into prelude/cxx/* making these rule definitions much more re-usable.

Reviewed By: milend

Differential Revision: D38124086

fbshipit-source-id: bdee406887d445c311a507f0f092c714e2e1876e