bxl: Migrate BxlResult::new to take IndexSet<ArtifactGroup>
Summary:
This is a small refactor to prepare for the next diff. We would like BxlResult to take on a vector of ArtifactGroups, not Artifact, so that we can materialize Artifacts and its associated ArtifactGroups.

There is also a minor type change to use IndexSet over HashSet as that is what will be returned by the StarlarkArtifactLike interface.

Reviewed By: bobyangyf

Differential Revision: D38374563

fbshipit-source-id: b9ed2770a330e0ccf2b5ae6cd539ed4e41fecb5c