buck2: event log: simplify Encoding / LogMode interaction
Summary:
We have an enum with 2 variants for mode and one with 4 variants for mode +
extension. It's a bit messy. Let's just have 1 enum for mode and 1 for
extension, and then a struct for the intersection of those things so that
not all places have to re-implement "what variants are protobuf?" and "what
variants are gzip?"

Differential Revision: D38530701

fbshipit-source-id: 0d5aef98f2e9043527c3ce66f76c78150661b692