Display a list of test failures and fatals in the summary
Summary:
In this diff we improve the output of test runs in buck2 by adding a list of failures to the summary as it happens in buck.
The main change was to change the data that the daemon passes. Instead of passing just counters now we pass counters with examples.

Reviewed By: krallin

Differential Revision: D38470388

fbshipit-source-id: 213de600be3c87b11a64601241f149cf0f69ad1e