handle relative symlinks inside symlinked folders on Windows
Summary:
Relative symlinks on Windows work not the way you'd expect. They are relative to real path on the filesystem and not the path you're accessing them with. This causes issues when we create symlinks from buck-out to repo (e.g. to source files).

With EdenFS `buck-out` is a symlink to some external directory and relative symlinks from `buck-out` to repo root are pointing to non existing location. To solve this we can check if relative path from link real path is indeed what we expect. It these paths are different, we can just create a symlink with absolute path.

Reviewed By: themarwhal

Differential Revision: D38365802

fbshipit-source-id: b2bf5a7647f5ef38a5c41c6748934f7e52d20a31