Improve the repr/Display for ctx and ctx.actions
Summary:
Sometimes these get printed to the user, e.g. on type error. The result is a bit of a mess, e.g. P520222844 that Josiah got. I think just saying these are the weird internal types is probably enough for users.

I considered whether to make ctx say the label it applies to, but usually that isn't interesting information (in errors it will be on the context), and sometimes a ctx doesn't have a label, so a bit more fiddly.

Reviewed By: nataliejameson, bobyangyf

Differential Revision: D38487086

fbshipit-source-id: f581fa5cfa20af08c025630950a2d3c5901e296c