Support for package_asset_libraries
Summary:
`prebuilt_native_library`s can specify they can be stored as assets, and so can `cxx_library`s. However, that only takes effect if `package_asset_libraries` is set to `True` on the APK itself.

This diff implements support for packaging such libraries as assets when that attribute is set to True.

The assets are packaged into `assets/lib`, per the test.

#forcetdhashing

Reviewed By: smeenai

Differential Revision: D38536388

fbshipit-source-id: 9fb4b24082a158039a2e660e98e990f3d15f03a1