cxx_library should also form the target graph for APKModuleGraph
Summary: `AndroidPackageableInfo` collects up all the targets in the target graph of an Android APK in order to "recreate" the target graph, which can then be used for determining which targets should go into each module. Previously, we weren't including `cxx` library targets in the target graph, now we do.

Reviewed By: milend

Differential Revision: D38387832

fbshipit-source-id: fd204dedc9f1a7c4b850fe80c79e195164fb3cb3