Account for aliases in android_app_modularity target graph
Summary:
The problem with aliases when building these graphs is that the alias just directly passes on the providers, but the label on the dep itself will be for the alias.

So say we have A (`java_library`) -> B (`alias`) -> C (`java_library`), prior to this diff the graph we build will have a dep from A to B, which isn't any use to us as B doesn't have a node in the graph (it just passes on C's information, and doesn't have any information of its own).

We don't want to add information to `alias` in any way, so the solution is to tie the deps info that we store to the target that produced it. This way, in the example above we get A to depend on C and everything works as expected.

Reviewed By: bchang7

Differential Revision: D38455547

fbshipit-source-id: 56b4491df47a78df8770b3bbb19b5e6ad9d726c7