Switch exec platforms to have Xcode constraints
Summary:
This adds two new macOS platforms to be used with Buck2's execution platforms. Those platforms are identical to the base platforms except with the addition of a relevant Xcode constraint.

This means that if a toolchain defines a specific Xcode constraint, all actions will be routed to the appropriate RE worker with the correct Xcode selected.

Furthermore, this allows us to support multiple Xcode versions on RE since we can have multiple execution platforms with different constraints.

(Note: this ignores all push blocking failures!)

Reviewed By: chatura-atapattu

Differential Revision: D38580329

fbshipit-source-id: d50f09c93a3bcb333d6efa2050a3ba4f1bfd0679