Update "test_fat_platforms" test to use RE Xcode 13.4
Summary:
It looks like 13.2 is no longer supported, so the action just fallsback to running locally.
This causes the test's remote-mac component to report it's running on "linux", failing
the test.

For some reason this test failure blamed to D37381834.  AFAICT, the 13.2 subplatform
has been gone for a while (the test actually fails on a rev before that diff if run now), so
I suspect that RE action caching has been keeping it going and that D37381834's
additional constraints caused the platform hashes (and therefore, the RE cache key)
to change, exposing this issue.

Reviewed By: krallin

Differential Revision: D38618702

fbshipit-source-id: 1b26bef73195aacda21a7a5f691cf3dcb4404c00