fix versioning when validating a version that removes a dirty
Summary:
Consider this:
verified: [1]
dirty [2]

and we verify v3

We currently stored this as
verified: [1, 3]
dirty: []

But we should store this as
verified: [1]
dirty: []

There's other lookup logic that scans the verified and dirties around that might be slightly confused by this representation, so better to fix it.

long term, I plan to change the data structures so this is all less confusing.

Reviewed By: jaspercb

Differential Revision: D38520409

fbshipit-source-id: fdf323bb45ff61648c3f63dacfc024b279687574