Mark some python_binary target compatible with fat platforms
Summary:
This changes python_binary to get its fat platform compatibility from
its toolchain. This makes the python toolchains mark their compatibility
with fat platforms based on whether its interpreter is in a defined list of
known compatible interpreters (the binaries we produce hardcode in the interpreter
path and so that path needs to be valid where they are executed).

Reviewed By: IanChilds

Differential Revision: D37477643

fbshipit-source-id: 26eab303acf22db8e1417c8829cc723386812cc9