Send Snapshot immediately on command start and on end
Summary:
milend noted it'd be great to see memory usage before and after a command executed. We store `max_rss` in a snapshot, so we could use that, but we currently send the first snapshot 1 second after command start. Let's send one immediately on command start, and also send another snapshot on drop.

This isn't a perfect proxy for command memory usage, since we might have concurrent commands running, but it's an okay one.

Reviewed By: krallin

Differential Revision: D38364975

fbshipit-source-id: e91d033633818878135aa310120557106186cde1