Buckconfigs must be empty in PlatformInfo provider
Summary:
We don't use buckconfigs from `PlatformInfo` providers, and they are always populated empty in `platform` rule and in transitions.

To do select based on buckconfig, we [fetch global buckconfig](https://fburl.com/code/cjmck7k5), but not buckconfig from target's configuration.

Add an assertion in code to make it clear.

And I think this is right: buckconfig should not be a part of configuration, we call it "legacy" for a reason. Is that true? If so, we should remove values from `PlatformInfo` provider.

Otherwise we should fill `values` and start using them.

Reviewed By: cjhopman

Differential Revision: D36339022

fbshipit-source-id: 8af1705a93edc4498e359f5fc2c23743ee78257a