Move sorted_index_set into buck2_common/src
Summary:
The plan is to use the SortedIndexSet wrapper in the artifacts work I'm doing in the next diff. The bulk of the change is to move the implementation up a directory so that it's under buck2_common.

I also added a Default trait implementation so that we can easily do Default::default().

Reviewed By: bobyangyf

Differential Revision: D38531469

fbshipit-source-id: e98a28d8a64973b90a97403d08c5f97e9517fdc3