Refactor in preparation for supporting modules
Summary: This introduces the concept of modules - we have a function to return the module that a target exists in, and then for the root module we do the same as before, if it's not the root module then we stored it as an asset. We keep a map of module to assets.

Reviewed By: smeenai

Differential Revision: D38616072

fbshipit-source-id: d540a0ba57ac7dc48f3a1b0db636abcbf3d0cf84