buck2: cxx: reorganize linkable graph provide provider creation
Summary:
I need to teorder C++ library providers so that the native_link_target is
available when we add the linkable graph provider. This will be useful to add
"hinted" roots for Omnibus since that is a bit of information that goes into
the linkable graph but needs the native_link_target (that is, libraries that
should _always_ be treated as roots).

Reviewed By: chatura-atapattu

Differential Revision: D38653802

fbshipit-source-id: 4af89390c0b4b651d16dc0ca9ae0e02f3b7d1bcd