Add a simple link_group_map rule
Summary:
## This stack

Link group mappings are the most computationally expensive component using link groups. Given a mapping, the computed mapping is the same, and can be shared between link groups. In buck1, this is cached in memory. In buck2, this must be propagated up by a child or recomputed.

Instead, if a `link_group_map` is it's own rule, it's mapping can be provided directly to all consumers, and we can perform the computation just once.

## This diff

Add a new rule that represents a link group map.

Reviewed By: d16r

Differential Revision: D38621721

fbshipit-source-id: ab18c8e0c5a75395c750795d54f07ffb0ddfd02f