Extract the attrs for link_group_map into a shared reusable method
Summary:
## This stack

Link group mappings are the most computationally expensive component using link groups. Given a mapping, the computed mapping is the same, and can be shared between link groups. In buck1, this is cached in memory. In buck2, this must be propagated up by a child or recomputed.

Instead, if a `link_group_map` is it's own rule, it's mapping can be provided directly to all consumers, and we can perform the computation just once.

## This diff

Extract the attrs definition so that it can be used in all referenced locations. Reduces duplication and makes it easy to update.

Reviewed By: d16r

Differential Revision: D38621722

fbshipit-source-id: 64988d2a3698d95bc4dfd34de5442f7707323beb