Handle having either a dependency or a map set as link_group_map
Summary:
## This stack

Link group mappings are the most computationally expensive component using link groups. Given a mapping, the computed mapping is the same, and can be shared between link groups. In buck1, this is cached in memory. In buck2, this must be propagated up by a child or recomputed.

Instead, if a `link_group_map` is it's own rule, it's mapping can be provided directly to all consumers, and we can perform the computation just once.

## This diff

In D38621720, we allowed setting `link_group_map` to be a target. Here, we add support to either using that target, or the map (as set in buck1 and existing link group usages).

Instead of returning an array of parsed `link_groups`, we return a `LinkGroupInfo` containing all necessary link group information.

We're then able to extract all necessary information from it.

Reviewed By: milend

Differential Revision: D38621724

fbshipit-source-id: caa0ab2f9c8febad69fd37bc61af31884bb3f09d