Support setting a target as a link group
Summary:
## This stack

Link group mappings are the most computationally expensive component using link groups. Given a mapping, the computed mapping is the same, and can be shared between link groups. In buck1, this is cached in memory. In buck2, this must be propagated up by a child or recomputed.

Instead, if a `link_group_map` is it's own rule, it's mapping can be provided directly to all consumers, and we can perform the computation just once.

## This diff

Allow setting a target as a link group. This will eventually allow us to set a target that is a `link_group_map` as the value for the `link_group_map` attribute on cxx_library, etc.

Reviewed By: d16r

Differential Revision: D38621720

fbshipit-source-id: e2e590cb9d2608db4dd77400061df8cef15171ed