Add groups to the LinkGroupInfo and ResourceGroupInfo providers
Summary:
## This stack

Link group mappings are the most computationally expensive component using link groups. Given a mapping, the computed mapping is the same, and can be shared between link groups. In buck1, this is cached in memory. In buck2, this must be propagated up by a child or recomputed.

Instead, if a `link_group_map` is it's own rule, it's mapping can be provided directly to all consumers, and we can perform the computation just once.

## This diff

In the next diff, to ease migration, we need to hold onto the parsed link groups. Add it to the relevant providers and set the value.

Reviewed By: milend

Differential Revision: D38621719

fbshipit-source-id: 244d1484fd48b1301454eb080ccdc31cb426fdc6