Compute the link group mapping/info for a link_group_map rule
Summary:
## This stack

Link group mappings are the most computationally expensive component using link groups. Given a mapping, the computed mapping is the same, and can be shared between link groups. In buck1, this is cached in memory. In buck2, this must be propagated up by a child or recomputed.

Instead, if a `link_group_map` is it's own rule, it's mapping can be provided directly to all consumers, and we can perform the computation just once.

## This diff

Compute the mapping for a link group rule. It's values are currently unused but will be soon.

Reviewed By: d16r

Differential Revision: D38621725

fbshipit-source-id: 8e8fc023843fb365c5acff8884d51a9fdb3a155e