Add a return type to make things more clear
Summary:
This let's us see exactly what the `_get_native_libs_and_assets` method returns.

The important thing for the purpose of using dynamic actions (which we need to support Voltron) is that it returns single artifacts, not lists of artifacts, so that we can declare in advance how many artifacts we expect the dynamic output to produce. (That's coming in a later diff, but explains the motivation at least).

Reviewed By: smeenai

Differential Revision: D38654444

fbshipit-source-id: f2c3800a3f4e4b77e8b87650ccbf5df1b785421e