Register a ResourceInfo in rust_library
Summary:
This makes it possible for a `rust_binary` or `cpp_binary` which depends transitively on a `rust_library` to access resources declared by the dependency.

Per {D38078866 (https://github.com/facebookincubator/buck2/commit/8433a0e8615bdb115f583efc002b0fb3f20c83db)}, we are using the exact same buck2 provider type for resources of a `cpp_library` as for resources of a `rust_library`, so there is no change needed in this diff to anything under buck2/prelude/cxx. The existing code in there will automatically pick up the resources that this diff makes `rust_library` provide.

Reviewed By: zertosh

Differential Revision: D38117486

fbshipit-source-id: a521924c86cd82968a6d332b6307344d4858d246