Notify users if no targets can be matched against a link group or mapping
Summary:
We currently don't inform users if there are no matches for a mapping or link group (e.g. due to a typo in the pattern string). This may lead to unexpected behavior for the user as the .so file will still be created but won't have the contents that they expect.

Based on offline discussion with Christy, we will just warn the user if there is a partial match (as parts of our default configs may not be relevant to their binary) but error out if there is no match for a link group.

Reviewed By: chatura-atapattu

Differential Revision: D38692036

fbshipit-source-id: 1360c38eba5eaabebfcc25ff1a46e344d808a3e4