Resolve frameworks when generating linker and compiler flags
Summary:
# Problems

There are currently two problems when expanding framework paths.

**Problem 1: Expanded at Declaration Site**

At present, if an `apple_library()` declares a framework like `$SDKROOT/Developer/Library/Frameworks/StoreKitTest.framework`, the expansion of `$SDKROOT` will happen *once* using the toolchain for the library.

This is **incorrect** because it makes an assumption that the `apple_binary()` which will use that `apple_library()` will also be linked with the same toolchain but that's not necessarily true.

**Problem 2: Incompatible with Pika Toolchain**

For `apple_toolchain()`, the platform and SDK paths are actually `artifact`s, so the current framework expansion logic is broken.

To be able to perform the expansion, we will need to use `cmd_args` to concat the platform/SDK paths (because they're `artifact`s). But once we use `cmd_args`, we cannot hash the resolved paths (because `cmd_args` is mutable and not hashable) and the hashing is required to be able to uniquify

https://www.internalfb.com/code/fbsource/[1f94ec5d2da33b6a982674c49262b8a84b927bf6]/fbcode/buck2/prelude/linking/link_info.bzl?lines=501-522

One possible way to workaround that limitation is to keep passing tuples of `artifact` and `string` everywhere and convert to `cmd_args` after uniquification.

That will become quite cumbersome.

# Solution

There's a simple solution that solves both problems: we perform framework expansion at the points where we need to pass resolved paths to the compiler/linker. Otherwise, we operate on unresolved framework paths (which are de-dupeable as before).

Reviewed By: d16r

Differential Revision: D38740522

fbshipit-source-id: 13d96667166cdffb4fd6bc5786790fb17bb8a67d