Native library rules now support Voltron modules
Summary:
This is the final piece for supporting Voltron modules for native library rules. We create a dynamic action that takes the file containing the Voltron target-to-module-mapping, and then use that to create our own mapping that we use to create all of the native libraries.

Also added a bunch of documentation to explain what is going on.

Reviewed By: smeenai

Differential Revision: D38654443

fbshipit-source-id: 072115bbda921d8856e07658f021117323290bd2