application_module_configs deps must use same configuration as primary platform of 'deps'
Summary: When we construct the APK module graph (in the next diff), we use both the `deps` field and the deps provided as config seeds. Since the `deps` field gets transitioned to configs representing the architectures we care about (and we use the first of those) we need to add a transition to match that configuration so our full deps graph is using the same configuration throughout.

Differential Revision: D38703096

fbshipit-source-id: 7c7a45a406e13df37980d3b56629b25237b1752c