Add a separate fbcode-based Python toolchain for the DEFAULT clause
Summary:
Since the beginning, one of fbcode's Python platforms has always sat in the
`DEFAULT` clause of the giant Python toolchain select which has caused a lot
of unexpected flows to depend on it (e.g. ARVR v2 TD) which in turn has
prevented making fbcode's tools' `target_compatible_with`s more strict.

This diff adds a new definition of an fbcode Python platform to slot in to the
`DEFAULT` clause so we can tune the real fbcode toolchains w/o breakages.

Reviewed By: milend

Differential Revision: D38742794

fbshipit-source-id: 49709c9d1335f59e855fcd43610c5cff4005707a