Correctly expand frameworks when using Pika toolchains
Summary:
# Problem

When using Pika toolchains, the platform and SDK paths are `artifact`s.

https://www.internalfb.com/code/fbsource/[5a7b46c362f958013549934fe4c219a1e4c3cddc]/fbcode/buck2/prelude/apple/apple_rules_impl.bzl?lines=105-106

https://www.internalfb.com/code/fbsource/[5a7b46c362f958013549934fe4c219a1e4c3cddc]/fbcode/buck2/prelude/apple/apple_toolchain.bzl?lines=6-7

This means that when we expand framework paths (e.g., `"$PLATFORM_DIR/Developer/Library/Frameworks/XCTest.framework"`), we cannot use simple string manipulation to concatenate an `artifact` and a string.

# Solution

We need to use `cmd_args` to perform the concatenation (using `delimiter = "/"` pattern).

#retry_on_user_failure #notimeout #retry_on_timeout

Reviewed By: chatura-atapattu

Differential Revision: D38741448

fbshipit-source-id: 518230c457d5aa025b4ba30d5ddacc8bdf17aeb0