Toolchain targets are now Xcode version namespaced
Summary:
Changes the names of all targets, so that we can have multiple toolchains supported. If the Xcode version was not in the name, we would get clashing targets, as they would have the same target name (e.g., `actool` target multiple times for each Xcode version).

New Xcode 14 toolchain added in D38787789.

Differential Revision: D38786504

fbshipit-source-id: f3f353cbae168b846199f544094a35d2a4b1d229