buck2_server crate
Summary:
Move parts of `cli` crate to `buck2_server` crate.

`buck2_server` crate name is maybe not ideal.

Initially I thought it should contain server part of `cli` crate, and while it is mostly the case, some parts of the client also rely in `buck2_server` crate. I can rename, but probably after the stack lands, otherwise rebase is a lot of work.

Current idea for the crate structure is this:

`cli` crate glues all pieces together, and the pieces are:
* `buck2_test` crate implements test
* `buck2_bxl` implements BXL
* `buck2_server` implements the rest of commands

With the top of the stack, `buck2_server` crate contains 8K LOC while `cli` contains 21K LOC. This looks significant, but it save only a few seconds of check or compilation time.

But it is a win, because `buck2_server` can be compiled concurrently with `buck2_bxl` and `buck2_test` crates. Dependency graph at the top of the stack:

{F761063570}

Reviewed By: bobyangyf

Differential Revision: D38686802

fbshipit-source-id: a3cbf7af571eaa0d05f801fd216885374b2de7a1