Add SharedTask span event type
Summary:
This commit adds a SharedTask span type that is used in the subsequent diff.

SharedTaskStart indicates when a command starts waiting on a Future that was spawned and is being executed in a separate command.

SharedTaskEnd indicates when the result has been computed ready for the command to use (unblocked).

Reviewed By: jaspercb

Differential Revision: D37660408

fbshipit-source-id: 8200c82c2dca670f410c0cd9e6888a3d4d85ea75