Pass dependencies into AST dumper plugin
Summary:
I'm getting ready to make the AST dumper add type info, including references to types on the classpath.

In order for that to be useful, the type reference information needs to include which Buck target defined the reference. The dumper can see the path of the JAR that defines a class, so if we provide a list of all the jars on the classpath and their Buck targets, it will be able to map back.

This diff updates Buck2 to pass the currently-building target name and the target names & JAR paths of everything on the classpath. For now it does the minimal update to the dumper plugin to accept those arguments and drop them on the floor; later diffs will do something interesting with them.

Reviewed By: IanChilds

Differential Revision: D38767901

fbshipit-source-id: 51a39b90f7fcf09482293c65d6c2d5d3b1582003